+++ app/page.js
@@ -1,11 +1,18 @@
-import Link from "next/link";
+import Images from "./ui/images";
+import Search from "./ui/search";
+import { ImagesSkeleton } from "./skeletons";
+import { Suspense } from "react";
 
-export default async function Home() {
-  
+export default async function Page({ searchParams }) {
+  const query = searchParams?.query || '';
+
   return (
-    <>
-      Home Page
-      <div><Link href='/photo-gallery'>To Gallery</Link></div>
-    </>
-  );
+    <div className='flex flex-col mx-auto my-0'>
+      <p className='text-center text-xl p-12'>Welcome to the Gallery</p>
+      <Search />
+      <Suspense key={searchParams.query} fallback={<ImagesSkeleton />}>
+        <Images query={query}/>
+      </Suspense>
+    </div>
+  )
 }
